load rules optionally

+++ server/src/main.ts
@@ -9,10 +9,12 @@ import { TwitterService } from './services/twitter.service';
   const twitterService = new TwitterService(configService, logger);
 
   logger.info('Starting up');
-  const rules = await twitterService.getRules();
-  await twitterService.deleteRules(rules);
-  await twitterService.setRules();
-  twitterService.flood();
+  if (configService.loadRules) {
+    const rules = await twitterService.getRules();
+    await twitterService.deleteRules(rules);
+    await twitterService.setRules();
+  }
+  // twitterService.flood();
 })().catch((e) => {
   console.log(JSON.stringify(e));
 });
